Cleanup and refactoring

+++ CBORTest/FastRandom.cs
@@ -40,7 +40,7 @@ namespace Test {
     /// <returns>A 32-bit signed integer.</returns>
     public int NextValue(int v) {
       if (v <= 0) {
- throw new ArgumentException("v (" + Convert.ToString((long)v, System.Globalization.CultureInfo.InvariantCulture) + ") is not greater than " + "0");
+ throw new ArgumentException("v (" + Convert.ToString((int)v, System.Globalization.CultureInfo.InvariantCulture) + ") is not greater than " + "0");
 }
       if (v <= 1) {
         return 0;
