+++ src/com/upokecenter/util/BitShiftAccumulator.java
@@ -54,7 +54,7 @@ at: http://upokecenter.com/d/
      */
     public void ShiftToDigits(FastInteger bits) {
       if (bits.signum() < 0) {
- throw new IllegalArgumentException("bits's sign (" + Long.toString((long)bits.signum()) + ") is less than " + "0");
+ throw new IllegalArgumentException("bits's sign (" + Integer.toString((int)bits.signum()) + ") is less than " + "0");
 }
       if (bits.CanFitInInt32()) {
         this.ShiftToDigitsInt(bits.AsInt32());
@@ -113,7 +113,7 @@ at: http://upokecenter.com/d/
       int lastDiscarded,
       int olderDiscarded) {
       if (bigint.signum() < 0) {
- throw new IllegalArgumentException("bigint's sign (" + Long.toString((long)bigint.signum()) + ") is less than " + "0");
+ throw new IllegalArgumentException("bigint's sign (" + Integer.toString((int)bigint.signum()) + ") is less than " + "0");
 }
       if (bigint.canFitInInt()) {
         this.isSmall = true;
@@ -128,7 +128,7 @@ at: http://upokecenter.com/d/
 
     public static BitShiftAccumulator FromInt32(int smallNumber) {
       if (smallNumber < 0) {
- throw new IllegalArgumentException("smallNumber (" + Long.toString((long)smallNumber) + ") is less than " + "0");
+ throw new IllegalArgumentException("smallNumber (" + Integer.toString((int)smallNumber) + ") is less than " + "0");
 }
       BitShiftAccumulator bsa = new BitShiftAccumulator(BigInteger.ZERO, 0, 0);
       bsa.shiftedSmall = smallNumber;
@@ -375,7 +375,7 @@ at: http://upokecenter.com/d/
      */
     public void ShiftToDigitsInt(int bits) {
       if (bits < 0) {
- throw new IllegalArgumentException("bits (" + Long.toString((long)bits) + ") is less than " + "0");
+ throw new IllegalArgumentException("bits (" + Integer.toString((int)bits) + ") is less than " + "0");
 }
       if (this.isSmall) {
         this.ShiftSmallToBits(bits);
